Add `Model` with shared code
