+++ src/vertex_ai_huggingface_inference_toolkit/utils/docker.py
@@ -18,7 +18,8 @@ def build_docker_image(
     python_version: str,
     framework: str,
     framework_version: str,
-    transformers_version: str,
+    transformers_version: Optional[str] = None,
+    diffusers_version: Optional[str] = None,
     cuda_version: Optional[str] = None,
     ubuntu_version: Optional[str] = None,
     extra_requirements: Optional[List[str]] = None,
@@ -35,6 +36,8 @@ def build_docker_image(
         framework_version: is the version of the provided framework as shown in PyPI.
         transformers_version: is the version of `transformers` to install, since the
             inference code will be run via `transformers`.
+        diffusers_version: is the version of `diffusers` to install, since the
+            inference code will be run via `diffusers`.
         cuda_version: is the version of CUDA to use, if planning to deploy the model
             within an instance with GPU acceleration. The CUDA versions to be provided
             need to be in the format of X.Y.Z, and available at https://hub.docker.com/r/nvidia/cuda/tags?page=1&name=-base-ubuntu
@@ -51,7 +54,12 @@ def build_docker_image(
 
     # The tag is set in advance, generated from the replacements of the `--build-args`
     _device_string = f"cu{cuda_version}" if cuda_version is not None else "cpu"
-    _tag = f"py{python_version}-{_device_string}-{framework}-{framework_version}-transformers-{transformers_version}"
+    _huggingface_string = (
+        f"transformers-{transformers_version}"
+        if transformers_version is not None
+        else f"diffusers-{diffusers_version}"
+    )
+    _tag = f"py{python_version}-{_device_string}-{framework}-{framework_version}-{_huggingface_string}"
 
     # The `_build_args` to be replaced in the `Dockerfile` when building it need to be
     # prepared in advance, to ensure the formatting and assignment is fine.
@@ -60,8 +68,11 @@ def build_docker_image(
         "PYTHON_VERSION": python_version,
         "FRAMEWORK": framework,
         "FRAMEWORK_VERSION": framework_version,
-        "TRANSFORMERS_VERSION": transformers_version,
     }
+    if transformers_version is not None:
+        _build_args["TRANSFORMERS_VERSION"] = transformers_version
+    if diffusers_version is not None:
+        _build_args["DIFFUSERS_VERSION"] = diffusers_version
     if cuda_version is not None:
         _build_args["CUDA_VERSION"] = cuda_version
         _dockerfile = "Dockerfile.gpu"
